refactor: move observed-workflow abstraction from command layer to core/observation
