+++ homeboy.json
@@ -684,8 +684,7 @@
         "thin_command_adapter::src/commands/observe.rs::ThinCommandAdapterViolation",
         "thin_command_adapter::src/commands/review/mod.rs::ThinCommandAdapterViolation",
         "thin_command_adapter::src/commands/runs/gh_actions.rs::ThinCommandAdapterViolation",
-        "thin_command_adapter::src/commands/runs/loop_sync.rs::ThinCommandAdapterViolation",
-        "thin_command_adapter::src/commands/utils/observed_workflow.rs::ThinCommandAdapterViolation"
+        "thin_command_adapter::src/commands/runs/loop_sync.rs::ThinCommandAdapterViolation"
       ],
       "metadata": {
         "alignment_score": 0.9021739363670349,
